+++ troff/troff.d/troff.h
@@ -53,7 +53,7 @@ extern	float		*zoomtab;
 
 extern	int		nchtab;
 extern	char		*chname;
-extern	short		*chtab;
+extern	size_t		*chtab;
 extern	int		c_endash;
 
 extern	int		kern;
